removed some unnecessary casts

+++ src/main/java/net/crazysnailboy/mods/enchantingtable/block/BlockEnderEnchantmentTable.java
@@ -41,12 +41,12 @@ public class BlockEnderEnchantmentTable extends net.minecraft.block.BlockEnchant
 		{
 			int j = rand.nextInt(2) * 2 - 1;
 			int k = rand.nextInt(2) * 2 - 1;
-			double d0 = (double)pos.getX() + 0.5D + 0.25D * (double)j;
-			double d1 = (double)((float)pos.getY() + rand.nextFloat());
-			double d2 = (double)pos.getZ() + 0.5D + 0.25D * (double)k;
-			double d3 = (double)(rand.nextFloat() * (float)j);
-			double d4 = ((double)rand.nextFloat() - 0.5D) * 0.125D;
-			double d5 = (double)(rand.nextFloat() * (float)k);
+			double d0 = pos.getX() + 0.5D + 0.25D * j;
+			double d1 = pos.getY() + rand.nextFloat();
+			double d2 = pos.getZ() + 0.5D + 0.25D * k;
+			double d3 = rand.nextFloat() * j;
+			double d4 = (rand.nextFloat() - 0.5D) * 0.125D;
+			double d5 = rand.nextFloat() * k;
 			worldIn.spawnParticle(EnumParticleTypes.PORTAL, d0, d1, d2, d3, d4, d5, new int[0]);
 		}
 	}
